+++ metadrive/component/pgblock/fork.py
@@ -7,7 +7,7 @@ from metadrive.component.pgblock.create_pg_block_utils import CreateRoadFrom, Cr
 from metadrive.component.pgblock.pg_block import PGBlock
 from metadrive.component.pgblock.ramp import Ramp
 from metadrive.component.lane.straight_lane import StraightLane
-from metadrive.component.road import Road
+from metadrive.component.road_network import Road
 from metadrive.constants import Decoration, LineType
 from metadrive.utils.scene_utils import check_lane_on_road
 from metadrive.utils.space import ParameterSpace, Parameter, BlockParameterSpace
