+++ metadrive/component/pgblock/ramp.py
@@ -6,7 +6,7 @@ from metadrive.component.pgblock.create_pg_block_utils import ExtendStraightLane
     create_bend_straight
 from metadrive.component.pgblock.pg_block import PGBlock
 from metadrive.component.lane.straight_lane import StraightLane
-from metadrive.component.road import Road
+from metadrive.component.road_network import Road
 from metadrive.constants import Decoration, LineType
 from metadrive.utils.scene_utils import check_lane_on_road
 from metadrive.utils.space import ParameterSpace, Parameter, BlockParameterSpace
