+++ metadrive/component/pgblock/tollgate.py
@@ -2,7 +2,7 @@ from metadrive.component.pgblock.bottleneck import PGBlock
 from metadrive.component.pgblock.create_pg_block_utils import CreateAdverseRoad, CreateRoadFrom, ExtendStraightLane
 from metadrive.component.pgblock.pg_block import PGBlockSocket
 from metadrive.component.buildings.tollgate_building import TollGateBuilding
-from metadrive.component.road import Road
+from metadrive.component.road_network import Road
 from metadrive.constants import LineType, LineColor
 from metadrive.engine.engine_utils import get_engine
 from metadrive.utils.space import ParameterSpace, Parameter, BlockParameterSpace
