--- metadrive/component/road/__init__.py
@@ -1,2 +1,0 @@
-from metadrive.component.road.road import Road, Route
-from metadrive.component.road.road_network import NodeRoadNetwork
