+++ metadrive/envs/marl_envs/marl_bottleneck.py
@@ -1,7 +1,7 @@
 from metadrive.component.map.pg_map import PGMap
 from metadrive.component.pgblock.bottleneck import Merge, Split
 from metadrive.component.pgblock.first_block import FirstPGBlock
-from metadrive.component.road import Road
+from metadrive.component.road_network import Road
 from metadrive.envs.marl_envs.marl_inout_roundabout import LidarStateObservationMARound
 from metadrive.envs.marl_envs.multi_agent_metadrive import MultiAgentMetaDrive
 from metadrive.manager.map_manager import MapManager
