+++ metadrive/envs/marl_envs/marl_inout_roundabout.py
@@ -5,7 +5,7 @@ import numpy as np
 from metadrive.component.map.pg_map import PGMap
 from metadrive.component.pgblock.first_block import FirstPGBlock
 from metadrive.component.pgblock.roundabout import Roundabout
-from metadrive.component.road import Road
+from metadrive.component.road_network import Road
 from metadrive.envs.marl_envs.multi_agent_metadrive import MultiAgentMetaDrive
 from metadrive.manager.map_manager import MapManager
 from metadrive.manager.spawn_manager import SpawnManager
