+++ metadrive/manager/object_manager.py
@@ -2,7 +2,7 @@ from metadrive.component.lane.abs_lane import AbstractLane
 from metadrive.component.pgblock.curve import Curve
 from metadrive.component.pgblock.ramp import InRampOnStraight, OutRampOnStraight
 from metadrive.component.pgblock.straight import Straight
-from metadrive.component.road import Road
+from metadrive.component.road_network import Road
 from metadrive.component.static_object.traffic_object import TrafficCone, TrafficWarning, TrafficBarrier
 from metadrive.engine.engine_utils import get_engine
 from metadrive.manager.base_manager import BaseManager
