+++ metadrive/tests/vis_block/vis_a_small_town.py
@@ -4,7 +4,7 @@ from metadrive.component.pgblock.roundabout import Roundabout
 from metadrive.component.pgblock.std_intersection import StdInterSection
 from metadrive.component.pgblock.straight import Straight
 from metadrive.component.pgblock.t_intersection import TInterSection
-from metadrive.component.road.road_network import NodeRoadNetwork
+from metadrive.component.road_network.node_road_network import NodeRoadNetwork
 from metadrive.engine.asset_loader import initialize_asset_loader
 from metadrive.tests.vis_block.vis_block_base import TestBlock
 from metadrive.utils.space import Parameter
