+++ metadrive/tests/vis_block/vis_big.py
@@ -1,5 +1,5 @@
 from metadrive.component.algorithm.BIG import BIG
-from metadrive.component.road.road_network import NodeRoadNetwork
+from metadrive.component.road_network.node_road_network import NodeRoadNetwork
 from metadrive.engine.asset_loader import initialize_asset_loader
 from metadrive.tests.vis_block.vis_block_base import TestBlock
