+++ metadrive/tests/vis_block/vis_in_ramp.py
@@ -1,7 +1,7 @@
 from metadrive.component.pgblock.first_block import FirstPGBlock
 from metadrive.component.pgblock.ramp import InRampOnStraight
 from metadrive.component.pgblock.straight import Straight
-from metadrive.component.road.road_network import NodeRoadNetwork
+from metadrive.component.road_network.node_road_network import NodeRoadNetwork
 from metadrive.engine.asset_loader import initialize_asset_loader
 from metadrive.tests.vis_block.vis_block_base import TestBlock
