+++ metadrive/tests/vis_block/vis_std_t_intersection.py
@@ -2,7 +2,7 @@ from metadrive.component.pgblock.curve import Curve
 from metadrive.component.pgblock.first_block import FirstPGBlock
 from metadrive.component.pgblock.std_t_intersection import StdTInterSection
 from metadrive.component.pgblock.straight import Straight
-from metadrive.component.road.road_network import NodeRoadNetwork
+from metadrive.component.road_network.node_road_network import NodeRoadNetwork
 from metadrive.tests.vis_block.vis_block_base import TestBlock
 
 if __name__ == "__main__":
